Auto-update IDN URLs from IANA page
Note that "all" no longer rebuilds IDN because I'm not confident that
doing so won't upset IANA for bulk transfers if done routinely.  We're
talking to IANA about polite bulk access.

An env var lets us override the source URL, useful for testing.

A follow-up commit will update the metadata/ sub-dir with the results of
this.

We need a per-URL cache of IDN rules because URLs get repeated for
different zones.